fix: Android 13 : enable send notification permission - EXO-70754
