+++ app/src/main/java/org/exoplatform/activity/WebViewActivity.java
@@ -20,8 +20,11 @@ package org.exoplatform.activity;
  *
  */
 
+import android.Manifest;
 import android.content.Intent;
 import android.content.SharedPreferences;
+import android.content.pm.PackageManager;
+import android.os.Build;
 import android.os.Bundle;
 import android.view.MenuItem;
 import android.view.View;
@@ -30,12 +33,20 @@ import android.widget.LinearLayout;
 import androidx.annotation.NonNull;
 import androidx.appcompat.app.AppCompatActivity;
 import androidx.appcompat.widget.Toolbar;
+import androidx.core.app.ActivityCompat;
+import androidx.core.content.ContextCompat;
+
+import com.crashlytics.android.Crashlytics;
+import com.google.android.gms.tasks.OnSuccessListener;
+import com.google.firebase.iid.FirebaseInstanceId;
+import com.google.firebase.iid.InstanceIdResult;
 
 import org.exoplatform.App;
 import org.exoplatform.R;
 import org.exoplatform.fragment.PlatformWebViewFragment;
 import org.exoplatform.fragment.WebViewFragment;
 import org.exoplatform.model.Server;
+import org.exoplatform.service.push.PushTokenStorage;
 import org.exoplatform.service.push.PushTokenSynchronizerLocator;
 import org.exoplatform.tool.ServerManagerImpl;
 import org.exoplatform.tool.ServerUtils;
@@ -43,6 +54,8 @@ import org.exoplatform.tool.ServerUtils;
 import java.net.MalformedURLException;
 import java.net.URL;
 import java.util.Date;
+
+import io.fabric.sdk.android.Fabric;
 
 /**
  * Activity that loads Platform into a web view
@@ -66,6 +79,25 @@ public class  WebViewActivity extends AppCompatActivity implements PlatformWebVi
   @Override
   public void onCreate(Bundle savedInstanceState) {
     super.onCreate(savedInstanceState);
+
+    // initialize Android with Push notifs
+    if (Build.VERSION.SDK_INT >= 33) {
+      if (ContextCompat.checkSelfPermission(WebViewActivity.this, Manifest.permission.POST_NOTIFICATIONS) != PackageManager.PERMISSION_GRANTED) {
+        ActivityCompat.requestPermissions(WebViewActivity.this, new String[]{Manifest.permission.POST_NOTIFICATIONS}, 101);
+      } else {
+        Fabric.with(this, new Crashlytics());
+        FirebaseInstanceId.getInstance().getInstanceId().addOnSuccessListener(new OnSuccessListener<InstanceIdResult>() {
+          @Override
+          public void onSuccess(InstanceIdResult instanceIdResult) {
+            String newToken = instanceIdResult.getToken();
+            PushTokenStorage.getInstance().setPushToken(newToken, getApplicationContext());
+            PushTokenSynchronizerLocator.getInstance().setTokenAndSync(newToken);
+          }
+        });
+      }
+    }
+
+
     setContentView(R.layout.activity_webview);
     checkConnectivity = new CheckConnectivity(WebViewActivity.this);
     // Toolbar hidden by default, visible on certain pages cf
