+++ build.gradle
@@ -55,6 +55,6 @@ task clean(type: Delete) {
 
 // This block encapsulate custom properties and make them available to all modules in the project.
 ext {
-    minSdkVersion = 28 // Android 9.0 Pie
-    compileSdkVersion = 31
+    minSdkVersion = 33 // Android 9.0 Pie
+    compileSdkVersion = 33
 }
